+++ CMSIS/DSP/Source/SVMFunctions/arm_svm_linear_predict_f16.c
@@ -151,7 +151,7 @@ void arm_svm_linear_predict_f16(
         acc0 = vfmaq_n_f16(acc0,acc2,*pDualCoef++);
         acc0 = vfmaq_n_f16(acc0,acc3,*pDualCoef++);
 
-        sum += vecAddAcrossF16Mve(acc0);
+        sum += (_Float16)vecAddAcrossF16Mve(acc0);
 
         pSrcA += numCols * 4;
         /*
@@ -221,7 +221,7 @@ void arm_svm_linear_predict_f16(
         acc0 = vmulq_n_f16(acc0,*pDualCoef++);
         acc0 = vfmaq_n_f16(acc0,acc1,*pDualCoef++);
 
-        sum += vecAddAcrossF16Mve(acc0);
+        sum += (_Float16)vecAddAcrossF16Mve(acc0);
 
         pSrcA += numCols * 2;
         row -= 2;
@@ -274,7 +274,7 @@ void arm_svm_linear_predict_f16(
         /*
          * Sum the partial parts
          */
-        sum += (_Float16)*pDualCoef++ * vecAddAcrossF16Mve(acc0);
+        sum += (_Float16)*pDualCoef++ * (_Float16)vecAddAcrossF16Mve(acc0);
 
     }
